Update return to main page
