+++ src/components/ui/NavMenu/NavMenu.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { FC } from 'react';
-import { usePathname } from 'next/navigation';
+
 import { Link } from 'react-scroll';
 
 import { navMenuData } from '@/data';
@@ -11,17 +11,6 @@ import { INavMenuProps } from './INavMenuProps.types';
 import s from './NavMenu.module.css';
 
 export const NavMenu: FC<INavMenuProps> = ({ onClick, className }) => {
-	const pathname = usePathname();
-
-	const handleClick = () => {
-		if (pathname === '/policy') {
-			window.location.href = '/';
-		}
-		else if (onClick) {
-			onClick();
-		}
-	};
-
 	return (
 		<nav className={`${className}`}>
 			<ul className='flex flex-col items-center gap-6 lg:flex-row'>
@@ -34,7 +23,7 @@ export const NavMenu: FC<INavMenuProps> = ({ onClick, className }) => {
 							to={section}
 							smooth={true}
 							duration={500}
-							onClick={() => handleClick()}
+							onClick={onClick}
 							href='#'
 							ignoreCancelEvents={true}
 							className={`relative px-0.5 pb-0.5 focus:outline-1 focus:outline-current ${s.active}`}
