+++ src/sections/Header/Header.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { FC, useState } from 'react';
+import { usePathname } from 'next/navigation';
 
 import { Logo } from '@/components/ui/Logo';
 import { NavMenu } from '@/components/ui/NavMenu';
@@ -12,20 +13,36 @@ import { headerData } from '@/data';
 
 import BurgerIcon from '@/../public/icons/menu.svg';
 
-export const Header:FC = () => {
+export const Header: FC = () => {
 	const [isOpened, setIsOpened] = useState(false);
+	const pathname = usePathname();
 
 	const toggleMenuOpen = () => {
 		setIsOpened(prev => !prev);
 		document.body.style.overflow = isOpened ? 'auto' : 'hidden';
 	};
 
+	const handleClick = () => {
+		if (pathname === '/policy') {
+			window.location.href = '/';
+		}
+		return;
+	};
+
 	return (
 		<header className='absolute left-0 top-0 z-20 w-full bg-transparent py-6'>
 			<div className='container flex items-center justify-between '>
 				<Logo />
-				<NavMenu className='notLg:hidden' />
-				<ButtonMain chapter='header' to='contacts' styles='notLg:hidden'>
+				<NavMenu
+					className='notLg:hidden'
+					onClick={() => handleClick()}
+				/>
+				<ButtonMain
+					chapter='header'
+					to='contacts'
+					styles='notLg:hidden'
+					onClick={() => handleClick()}
+				>
 					{headerData.text}
 				</ButtonMain>
 				<ButtonLess
